Update gitignore and remove Parcelize from ProgressData

+++ app/src/main/java/com/stefan/simplebackup/utils/file/FileUtil.kt
@@ -72,6 +72,7 @@ object FileUtil {
         if (!dir.isDirectory) throw IOException("File must be verified to be directory beforehand")
         withContext(ioDispatcher) {
             dir.walkTopDown().forEach { dirFile ->
+                if (dirFile == dir) return@forEach
                 deleteFile(dirFile.absolutePath)
             }
         }
